camate-backend/requirements.txt
Django==4.2.0
dj-database-url==2.1.0
bcrypt==4.1.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
boto3==1.34.0
python-decouple==3.8
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
requests==2.31.0
gunicorn==21.2.0